Ensure 32-bit for pixel data. Fixes crash on x86_64.
git-svn-id: https://svn.aros.org/svn/aros/trunk@54850 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ contrib/Games/spout/spout.c
@@ -2,6 +2,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+#include <stdint.h>
 
 #ifdef HAVE_WORDEXP_H
 #include <wordexp.h>
@@ -341,19 +342,19 @@ pceAppProc (int cnt)
 
 	      if (upperLine == 111 && height > 0)
 		{
-		  unsigned long *pL;
-		  pL = (unsigned long *) (vbuff2 + 128 * 108 + 4);
-		  while (pL < (unsigned long *) (vbuff2 + 128 * 109 - 4))
+		  uint32_t *pL;
+		  pL = (uint32_t *) (vbuff2 + 128 * 108 + 4);
+		  while (pL < (uint32_t *) (vbuff2 + 128 * 109 - 4))
 		    {
 		      *pL++ = 0;
 		    }
 		  pL += 2;
-		  while (pL < (unsigned long *) (vbuff2 + 128 * 110 - 4))
+		  while (pL < (uint32_t *) (vbuff2 + 128 * 110 - 4))
 		    {
 		      *pL++ = 0xd3d3d3d3;
 		    }
 		  pL += 2;
-		  while (pL < (unsigned long *) (vbuff2 + 128 * 111 - 4))
+		  while (pL < (uint32_t *) (vbuff2 + 128 * 111 - 4))
 		    {
 		      *pL++ = 0;
 		    }
@@ -414,7 +415,7 @@ pceAppProc (int cnt)
 
 	  if ((upperLine & 31) == 0)
 	    {
-	      unsigned long *pL;
+	      uint32_t *pL;
 	      pceLCDSetBuffer (vbuff2 + ((upperLine - 24) & 127) * 128);
 	      pceFontSetBkColor (0);
 
@@ -422,7 +423,7 @@ pceAppProc (int cnt)
 		{
 		case 0:
 		  pL =
-		    (unsigned long *) (vbuff2 + 12 +
+		    (uint32_t *) (vbuff2 + 12 +
 				       ((upperLine - 24) & 127) * 128);
 		  for (i = 0; i < 16; i++)
 		    {
@@ -749,14 +750,14 @@ pceAppProc (int cnt)
   dispPos = upperLine;
 
   {
-    unsigned long *pL, *pL2, *pLe;
-    pL = (unsigned long *) (vbuff + 2 * 128);
-    pL2 = (unsigned long *) (vbuff2 + dispPos * 128);
+    uint32_t *pL, *pL2, *pLe;
+    pL = (uint32_t *) (vbuff + 2 * 128);
+    pL2 = (uint32_t *) (vbuff2 + dispPos * 128);
 
     pLe = pL2 + 128 * 78 / 4;
-    if (pLe > (unsigned long *) (vbuff2 + 128 * 128))
+    if (pLe > (uint32_t *) (vbuff2 + 128 * 128))
       {
-	pLe = (unsigned long *) (vbuff2 + 128 * 128);
+	pLe = (uint32_t *) (vbuff2 + 128 * 128);
       }
 
     while (pL2 < pLe)
@@ -766,8 +767,8 @@ pceAppProc (int cnt)
 	pL2++;
       }
 
-    pL2 = (unsigned long *) (vbuff2);
-    while (pL < (unsigned long *) (vbuff + 128 * (78 + 2)))
+    pL2 = (uint32_t *) (vbuff2);
+    while (pL < (uint32_t *) (vbuff + 128 * (78 + 2)))
       {
 	*pL = *pL2 & 0x03030303;
 	pL++;
